allow more headers to be unsigned in V4 requests
This change allows more SigV4 request headers to be excluded from
signature calculation. Previously, all headers with the "X-Amz-" prefix
were required to be included in signature calculation, and if any were
missing from the list of signed headers, an error would be returned.
Now, only the "Host" and "X-Amz-Date" headers are required to be
signed.